chore(core): hoist cloneValue from builder.ts into utils.ts
